Dispatch INSERT_OBJECTS event only once in ObjectPersister
