+++ Doctrine/AbstractProvider.php
@@ -93,13 +93,6 @@ abstract class AbstractProvider extends BaseAbstractProvider
                 $objects = $this->filterObjects($options, $objects);
 
                 if (!empty($objects)) {
-
-                    if ($this->dispatcher) {
-                        $event = new PersistingEvent($objects);
-                        $this->dispatcher->dispatch(PersistingEvent::INSERT_OBJECTS, $event);
-                        $objects = $event->getObjects();
-                    }
-
                     $this->objectPersister->insertMany($objects);
                 }
             } catch (BulkResponseException $e) {
