+++ Propel/Provider.php
@@ -30,13 +30,6 @@ class Provider extends AbstractProvider
                 ->getArrayCopy();
             $objects = $this->filterObjects($options, $objects);
             if (!empty($objects)) {
-
-                if ($this->dispatcher) {
-                    $event = new PersistingEvent($objects);
-                    $this->dispatcher->dispatch(PersistingEvent::INSERT_OBJECTS, $event);
-                    $objects = $event->getObjects();
-                }
-
                 $this->objectPersister->insertMany($objects);
             }
